+++ src/AgentFrameworkToolkit.Google/GoogleAgentFactory.cs
@@ -63,7 +63,7 @@ public class GoogleAgentFactory
     {
         IChatClient client = GetClient(options.Model);
 
-        AIAgent innerAgent = new ChatClientAgent(client, CreateChatClientAgentOptions(options));
+        AIAgent innerAgent = new ChatClientAgent(client, CreateChatClientAgentOptions(options), options.LoggerFactory, options.Services);
 
         // ReSharper disable once ConvertIfStatementToReturnStatement
         if (options.RawToolCallDetails != null)