+++ src/AgentFrameworkToolkit.OpenAI/OpenAIAgentFactory.cs
@@ -116,7 +116,11 @@ public class OpenAIAgentFactory
 
         AIAgent innerAgent = client
             .GetChatClient(options.Model)
-            .CreateAIAgent(chatClientAgentOptions);
+            .CreateAIAgent(
+                options: chatClientAgentOptions,
+                services: options.Services,
+                loggerFactory: options.LoggerFactory,
+                clientFactory: options.ClientFactory);
 
         // ReSharper disable once ConvertIfStatementToReturnStatement
         if (options.RawToolCallDetails != null)
@@ -140,7 +144,11 @@ public class OpenAIAgentFactory
 
         AIAgent innerAgent = client
             .GetChatClient(options.Model)
-            .CreateAIAgent(chatClientAgentOptions);
+            .CreateAIAgent(
+                options: chatClientAgentOptions,
+                services: options.Services,
+                loggerFactory: options.LoggerFactory,
+                clientFactory: options.ClientFactory);
 
         // ReSharper disable once ConvertIfStatementToReturnStatement
         if (options.RawToolCallDetails != null)
@@ -150,7 +158,7 @@ public class OpenAIAgentFactory
 
         return new OpenAIAgent(innerAgent);
     }
-    
+
     private ChatClientAgentOptions CreateChatClientAgentOptions(OpenAIAgentOptions options, OpenAIAgentOptionsForChatClientWithoutReasoning? chatClientWithoutReasoning, OpenAIAgentOptionsForResponseApiWithoutReasoning? responseWithoutReasoning, OpenAIAgentOptionsForResponseApiWithReasoning? responsesApiReasoningOptions, OpenAIAgentOptionsForChatClientWithReasoning? chatClientReasoningOptions)
     {
         bool anyOptionsSet = false;