+++ src/Samples/Providers/AzureOpenAI.cs
@@ -1,9 +1,8 @@
 ﻿using AgentFrameworkToolkit.AzureOpenAI;
 using AgentFrameworkToolkit.OpenAI;
-using Azure.Identity;
+using AgentFrameworkToolkit.Tools;
 using Microsoft.Agents.AI;
 using Microsoft.Extensions.AI;
-using OpenAI.Chat;
 
 #pragma warning disable OPENAI001
 
@@ -11,6 +10,12 @@ namespace Samples.Providers;
 
 public static class AzureOpenAI
 {
+    [AITool]
+    static string GetWeather()
+    {
+        return "Sunny";
+    }
+
     public static async Task RunAsync()
     {
         Configuration configuration = ConfigurationManager.GetConfiguration();
@@ -20,22 +25,20 @@ public static class AzureOpenAI
             ApiKey = configuration.AzureOpenAiKey
         };
 
-        AzureOpenAIEmbeddingFactory azureOpenAIEmbeddingFactory = new(connection);
-
-        IEmbeddingGenerator<string, Embedding<float>> embeddingGenerator = azureOpenAIEmbeddingFactory.GetEmbeddingGenerator("text-embedding-3-small");
-        GeneratedEmbeddings<Embedding<float>> generatedEmbeddings = await embeddingGenerator.GenerateAsync(["Hello"]);
-
         AzureOpenAIAgentFactory factory = new(connection);
 
+        AIToolsFactory aiToolsFactory = new();
+
+        IList<AITool> aiTools = aiToolsFactory.GetTools(typeof(AzureOpenAI));
         AzureOpenAIAgent agent = factory.CreateAgent(new OpenAIAgentOptionsForChatClientWithReasoning
         {
             Model = OpenAIChatModels.Gpt41Mini,
-            ReasoningEffort = ChatReasoningEffortLevel.Low,
             Instructions = "Speak like a pirate",
-            RawHttpCallDetails = details => Console.WriteLine(details.RequestData)
+            RawHttpCallDetails = details => Console.WriteLine(details.RequestData),
+            Tools = aiTools
         });
 
-        AgentRunResponse response = await agent.RunAsync<string>("Hello");
+        AgentRunResponse response = await agent.RunAsync<string>("How is the weather?");
         Console.WriteLine(response);
     }
 }