+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog - Agent Framework Toolkit
 
+## Version 1.0.0-preview.251217.1
+- Added `AIToolsFactory` to make it easier to create and define AI Tools
+- Added support for `IServiceProvider`, `ILoggerFactory` and `ClientFactory` in Agent Creation
+
 ## Version 1.0.0-preview.251215.1
 - Added option to get Raw Client from the various Connection Objects (Except for the Google Connection as switch to the new official Google Nuget is expected soon)
 - OpenRouter and XAI Connection are now inherited from OpenAI Connections